refactor: move update package to updater, and check to be NoUpdater
