+++ internal/updater/config.go
@@ -1,8 +1,8 @@
-package update
+package updater
 
 // repository is the GitHub repository to query for releases, set via ldflags
 // at build time. Empty in dev builds so go run / go build skip the check.
-// Example: -ldflags "-X github.com/signoz/foundry/internal/update.repository=SigNoz/foundry".
+// Example: -ldflags "-X github.com/signoz/foundry/internal/updater.repository=SigNoz/foundry".
 var repository string = "<unset>"
 
 // Config holds update notifier configuration.
